Move constants and fixtures out of test file
