+++ istio/tests/conftest.py
@@ -3,9 +3,12 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import os
 
+import mock
 import pytest
 import requests
+from requests.exceptions import HTTPError
 
+from datadog_checks.base.utils.common import ensure_unicode
 from datadog_checks.dev.conditions import CheckEndpoints
 from datadog_checks.dev.kube_port_forward import port_forward
 from datadog_checks.dev.terraform import terraform_run
@@ -49,3 +52,80 @@ def dd_environment():
                 # Generate some traffic
                 requests.get(page)
             yield instance
+
+
+class MockResponse:
+    """
+    MockResponse is used to simulate the object requests.Response commonly returned by requests.get
+    """
+
+    def __init__(self, content, content_type, status=200):
+        self.content = content if isinstance(content, list) else [content]
+        self.headers = {'Content-Type': content_type}
+        self.status = status
+        self.encoding = 'utf-8'
+
+    def iter_lines(self, **_):
+        content = self.content.pop(0)
+        for elt in content.split("\n"):
+            yield ensure_unicode(elt)
+
+    def raise_for_status(self):
+        if self.status != 200:
+            raise HTTPError('Not 200 Client Error')
+
+    def close(self):
+        pass
+
+
+@pytest.fixture
+def mesh_mixture_fixture():
+    mesh_file_path = os.path.join(os.path.dirname(__file__), 'fixtures', 'istio', 'mesh.txt')
+    mixer_file_path = os.path.join(os.path.dirname(__file__), 'fixtures', 'istio', 'mixer.txt')
+    responses = []
+    with open(mesh_file_path, 'r') as f:
+        responses.append(f.read())
+    with open(mixer_file_path, 'r') as f:
+        responses.append(f.read())
+
+    with mock.patch('requests.get', return_value=MockResponse(responses, 'text/plain'), __name__="get"):
+        yield
+
+
+@pytest.fixture
+def new_mesh_mixture_fixture():
+    files = ['mesh.txt', 'mixer.txt', 'pilot.txt', 'galley.txt', 'citadel.txt']
+    responses = []
+    for filename in files:
+        file_path = os.path.join(os.path.dirname(__file__), 'fixtures', '1.1', filename)
+        with open(file_path, 'r') as f:
+            responses.append(f.read())
+
+    with mock.patch('requests.get', return_value=MockResponse(responses, 'text/plain'), __name__="get"):
+        yield
+
+
+@pytest.fixture
+def new_pilot_fixture():
+    files = ['pilot.txt']
+    responses = []
+    for filename in files:
+        file_path = os.path.join(os.path.dirname(__file__), 'fixtures', '1.1', filename)
+        with open(file_path, 'r') as f:
+            responses.append(f.read())
+
+    with mock.patch('requests.get', return_value=MockResponse(responses, 'text/plain'), __name__="get"):
+        yield
+
+
+@pytest.fixture
+def new_galley_fixture():
+    files = ['galley.txt']
+    responses = []
+    for filename in files:
+        file_path = os.path.join(os.path.dirname(__file__), 'fixtures', '1.1', filename)
+        with open(file_path, 'r') as f:
+            responses.append(f.read())
+
+    with mock.patch('requests.get', return_value=MockResponse(responses, 'text/plain'), __name__="get"):
+        yield
